Fix Local Goal Calculation Error
Fix a bug where the local goal is set to the final goal
when the final goal is within the voronoi cell but outside of the
buffered voronoi cell, causing possible collisions with other robots.
Fixed by setting the local goal to the final goal only if it is inside
the buffered voronoi cell.